correct scaling on any monitor and soft render resolution

+++ Engine/Engine/Src/render/ScreenTexture.h
@@ -11,8 +11,8 @@ private:
 public:
 	~ScreenTexture();
 
-	static void Generate(const uint8_t* buffer);
-	static void Update(const uint8_t* buffer);
+	static void Generate(const uint32_t* buffer);
+	static void Update(const uint32_t* buffer);
 	static void Bind(unsigned int slot = 0);
 	static void Unbind();
 };
